Refactor `FileGenerator`.
